ik code works with non-renamed joints..

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/MotionCaptureLoader/ik/bvh_inverseKinematics.h
@@ -76,7 +76,7 @@ struct ikProblem
 
 
 //Temporary call that allows outside control..
-float approximateBodyFromMotionBufferUsingInverseKinematics(
+int approximateBodyFromMotionBufferUsingInverseKinematics(
                                          struct BVH_MotionCapture * mc,
                                          struct simpleRenderer *renderer,
                                          struct MotionBuffer * solution,
